refactor(remote-config): move UiConfigProvider to its own uiconfig package
UiConfigProvider lived in common.remoteconfig, but it's ui_config-feature-
specific, not a generic remote-config primitive — same reasoning that already
put WorkflowsConfigProvider in common.workflows rather than common.remoteconfig.
Moves UiConfigProvider + its test to common.uiconfig, no behavior change.
